# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Self-registration; account waits for admin approval",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "X-Client",
            "in": "header",
            "description": "web (default, cookie refresh) | mobile (refresh in body)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password-reset-requests": {
      "post": {
        "operationId": "AuthController_passwordResetRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Files a manual reset request for admins; always returns ok",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/stores": {
      "get": {
        "operationId": "PublicController_stores",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/stores/{id}": {
      "get": {
        "operationId": "PublicController_store",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/stores/{id}/categories": {
      "get": {
        "operationId": "PublicController_storeCategories",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/stores/{id}/products": {
      "get": {
        "operationId": "PublicController_storeProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "barcode",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "in_stock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/stores/{id}/offers": {
      "get": {
        "operationId": "PublicController_storeOffers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
    "/api/products/by-barcode": {
      "get": {
        "operationId": "PublicController_productsByBarcode",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 3,
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public"
        ]
      }
    },
...
    "/api/catalog/categories": {
      "get": {
        "operationId": "CatalogController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/catalog/items": {
      "get": {
        "operationId": "CatalogController_items",
        "parameters": [
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "exclusiveMinimum": false,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "name",
              "type": "string",
              "enum": [
                "name",
                "newest",
                "usage"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasl API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^(77|78|71|73)\\d{7}$"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 72
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "user_type": {
            "type": "string",
            "enum": [
              "customer",
              "merchant"
            ]
          },
          "city": {
            "type": "string",
            "maxLength": 80
          },
          "district": {
            "type": "string",
            "maxLength": 80
          },
          "address": {
            "type": "string",
            "maxLength": 200
          },
          "business_name": {
            "type": "string",
            "maxLength": 120
          },
          "business_category_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "phone",
          "password",
          "name",
          "user_type"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^(77|78|71|73)\\d{7}$"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 72
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 20
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 72
          },
          "new_password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 72
          }
        },
        "required": [
          "new_password"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^(77|78|71|73)\\d{7}$"
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "phone"
        ]
      },
...
    }
  }
}